feat: WIP (1) Example of var support from custom local go parser
This is a hacky poc to show how we can intercept the cli for TF files. We first used gopherjs to compile our local custom go parser into js, and we exported that function (parseModule) for use in the CLI. Then, we get the parsedFiles and we override the ones that are of TF to include the big JSON output of the parser, which contains the var deref.
There is duplication at the moment and there are a lot of things to consider, such as different structure, different flow, FFs, etc.