feat: WIP (2) Example of var support from custom local go parser
CFG-1323

Here we call the parser for each TF file and do the dereferencing there. This would work for just .tf files in this poc, but we can extend to cover .tfvars etc.

This is a hacky poc to show how we can intercept the cli for TF files. We first used gopherjs to compile our local custom go parser into js, and we exported that function (parseModule) for use in the CLI. Then, when we find a TF file we want to parse, we call the exported parser function on each file and we override the dereferenced jsonContent. We can easily put this behind a feature flag, etc to have control of what files we parse.